+++ audio_codec/libcook/helix_types.h
@@ -163,7 +163,7 @@ extern "C" {
 #elif defined(_VXWORKS)
     typedef unsigned int UINT32;
 #else
-    typedef unsigned int UINT32;
+    typedef unsigned long UINT32;
 #endif
 
 #endif /* #ifndef UINT32 */
